core/skills/deepcamera/skills/channels/line/requirements.txt
# LINE Channel
line-bot-sdk>=3.5.0
